requirements.txt
pandas
openpyxl
pydantic
openai
python-dotenv
volcengine-python-sdk[ark]
streamlit
langgraph
langchain
langchain-openai
jieba
sentence-transformers
numpy
psycopg2-binary
PyJWT
cryptography
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
